Downloader // use the "updated" date for translations as filemtime(). Ensure that updated zip files are downloaded again.
GlotPress API allows to keep updating translations for an already released version which is represented by "updated"-field in REST API. It is possible, that an existing plugin, theme or WP core can have release as version 1.0, but the translation will be updated after that release was done. Currently, our file caching for translation zip files was only storing the zip files, which contained the release version in file name. Downloaded zip files were on installation marked as "CACHED", because the version did not change, but in reality the "updated"-date changed. With this commit we will ensure that the downloaded zip file will get the "updated"-date as file time to allow better comparison and detection on next installation for "cached" or "not cached".