Small improvement to schema
api.names api.types, as well as directories.names and directories.types are *not* required.

However, if they are present as empty object a configuration validation error is raised. That is due to JSON encoding converting JSON '{}' to `[]` in PHP, which is then seen as an array, not an object.

Considering those props are not required, we should accept empty value. That is done offering empty array as an alternative to object in the schema definition.