Make use of HttpDownloader and JsonFile for downloading the TranslatePackage translations.
Previously the JSON data was downloaded via file_get_contents() which did not allow to make use of the stored credentials in composer config file. This implementation will make use of the Composer HttpDownloader which uses internally the Composer Config and the stored credentials to send alongside with the Request API Tokens or Basic Auth credentials.